+++ llm_guard/output_scanners/ban_topics.py
@@ -1,14 +1,11 @@
-import logging
 from typing import List
 
-from transformers import pipeline
-
-from llm_guard.util import device
+from llm_guard.util import device, lazy_load_dep, logger
 
 from .base import Scanner
 
-log = logging.getLogger(__name__)
 _model_path = "MoritzLaurer/mDeBERTa-v3-base-xnli-multilingual-nli-2mil7"
+transformers = lazy_load_dep("transformers")
 
 
 class BanTopics(Scanner):
@@ -35,12 +32,12 @@ class BanTopics(Scanner):
 
         self._topics = topics
         self._threshold = threshold
-        self._classifier = pipeline(
+        self._classifier = transformers.pipeline(
             "zero-shot-classification",
             model=_model_path,
-            device=device,
+            device=device(),
         )
-        log.debug(f"Initialized model {_model_path} on device {device}")
+        logger.debug(f"Initialized model {_model_path} on device {device()}")
 
     def scan(self, prompt: str, output: str) -> (str, bool, float):
         if output.strip() == "":
@@ -53,13 +50,13 @@ class BanTopics(Scanner):
 
         max_score = round(max(classifier_output["scores"]) if classifier_output["scores"] else 0, 2)
         if max_score > self._threshold:
-            log.warning(
+            logger.warning(
                 f"Topics detected for the prompt {classifier_output['labels']} with scores: {classifier_output['scores']}"
             )
 
             return output, False, max_score
 
-        log.debug(
+        logger.debug(
             f"No banned topics detected ({classifier_output['labels']}, scores: {classifier_output['scores']})"
         )